backend/ai/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Google Gemini API
google-genai>=0.3.0

# Supabase deps require httpx<0.28
httpx[http2]>=0.26,<0.28

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
